feat: Update createElement function to support data attributes
Extended the functionality of the createElement function to support data attributes in addition to regular attributes. Adjusted the implementation to set data attributes on the created HTML element using the dataset property. This enhancement provides users with more flexibility when creating HTML elements, allowing them to include custom data attributes as needed.