chore: Update dependencies and remove unnecessary devDependencies
Updated the dependencies in package.json to reflect the latest versions of TypeScript and Vite, bumping TypeScript to version 5.4.3 and Vite to version 5.2.2. Removed the "@types/node" and "prettier" devDependencies, as they are not needed for the project. These changes ensure that the project is up-to-date with the latest development tools and removes unnecessary dependencies, improving overall maintainability and reducing package size.